Cria componente CustomMarker
